+++ src/main/java/io/github/astrapi69/regex/RegExExtensions.java
@@ -18,12 +18,12 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.regex;
+package io.github.astrapi69.regex;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import de.alpharogroup.string.StringExtensions;
+import io.github.astrapi69.string.StringExtensions;
 
 /**
  * Utility class for the use with regular expressions.