+++ src/test/java/io/github/astrapi69/string/SplitStringExtensionsTest.java
@@ -18,7 +18,7 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.string;
+package io.github.astrapi69.string;
 
 import io.github.astrapi69.collections.array.ArrayFactory;
 import io.github.astrapi69.collections.map.MapFactory;